backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
facenet-pytorch>=2.5.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
python-multipart>=0.0.6
pydantic>=2.0.0
matplotlib>=3.7.0
scipy>=1.11.0




